Pipeline house cleaning
