+++ frontend/packages/dev-console/src/components/pipelines/detail-page-tabs/__tests__/pipeline-form-validation-utils.spec.ts
