+++ frontend/packages/dev-console/src/components/pipelines/detail-page-tabs/index.ts
@@ -4,3 +4,5 @@ export { default as PipelineParametersForm } from './PipelineParametersForm';
 export { default as PipelineResources } from './PipelineResources';
 export { default as PipelineResourcesForm } from './PipelineResourcesForm';
 export { default as PipelineRuns } from './PipelineRuns';
+export { default as PipelineForm } from './PipelineForm';
+export * from './pipelineForm-validation-utils';
