+++ frontend/packages/dev-console/src/components/pipelines/detail-page-tabs/pipelineForm-validation-utils.ts
