+++ frontend/packages/dev-console/src/components/pipelines/modals/common/PipelineSecretSection.scss
