+++ frontend/packages/dev-console/src/components/pipelines/modals/common/SecretForm.scss
