+++ frontend/packages/dev-console/src/components/pipelines/modals/common/SecretsList.scss
