+++ frontend/packages/dev-console/src/components/pipelines/modals/common/SecretsList.tsx
@@ -10,8 +10,8 @@ import { SecretModel, ServiceAccountModel } from '@console/internal/models';
 import { SecretType } from '@console/internal/components/secrets/create-secret';
 import { SecondaryStatus } from '@console/shared';
 import { SecretKind } from '@console/internal/module/k8s';
-import { ServiceAccountType } from '../../../utils/pipeline-utils';
-import { PIPELINE_SERVICE_ACCOUNT } from '../const';
+import { ServiceAccountType } from '../../../../utils/pipeline-utils';
+import { PIPELINE_SERVICE_ACCOUNT } from '../../const';
 import './SecretsList.scss';
 
 type SecretsProps = {
