+++ frontend/packages/dev-console/src/components/pipelines/modals/common/__mocks__/PipelineResourceDropdownField.tsx
