+++ frontend/packages/dev-console/src/components/pipelines/modals/index.ts
@@ -1,1 +1,2 @@
 export { default as removeTriggerModal } from './triggers/RemoveTriggerModal';
+export { default as startPipelineModal } from './start-pipeline/StartPipelineModal';
