+++ frontend/packages/dev-console/src/components/pipelines/pipeline-resource/__mocks__/pipeline-resource-mock.ts
