+++ frontend/packages/dev-console/src/components/pipelines/pipeline-resource/__tests__/pipelineResource-validation-utils.spec.ts
@@ -1,6 +1,6 @@
 import { cloneDeep } from 'lodash';
-import { validationSchema } from '../pipeline-resource/pipelineResource-validation-utils';
-import { getDefinedObj } from '../pipeline-resource/pipelineResource-utils';
+import { validationSchema } from '../pipelineResource-validation-utils';
+import { getDefinedObj } from '../pipelineResource-utils';
 import { mockPipelineResourceData } from '../__mocks__/pipeline-resource-mock';
 
 describe('Validation Pipeline Resource', () => {
